refactor ApiMock usage: create BaseApiMock
